Move mapif functions from int_mercenary.c to mapif.c
